% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{calculate_optimal_tiles}
\alias{calculate_optimal_tiles}
\title{Calculate Optimal Number of Tiles}
\usage{
calculate_optimal_tiles(num_codons)
}
\arguments{
\item{num_codons}{An integer representing the total number of codons.}
}
\value{
A list containing two elements: \code{optimal_tiles}, the number of tiles
that minimize the oligo cost, and \code{length_tiles_global}, the average
length of these tiles.
}
\description{
This function determines the optimal number of tiles for a given number of codons.
It calculates the oligo cost for a range of tile numbers and finds the number of
tiles that minimizes this cost. It also computes the global length of the tiles.
}
\examples{
# Calculate optimal tiles for 100 codons
calculate_optimal_tiles(100)

}
