% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{execute_and_plot}
\alias{execute_and_plot}
\title{Execute the Optimization Process and Plot Results}
\usage{
execute_and_plot(target_gene = RAD_27, max_iter = 30, scan_rate = 7)
}
\arguments{
\item{target_gene}{The target gene for assembly (default: Rad27),
must contain one codon before and after ORF.}

\item{max_iter}{The maximum number of iterations for the optimization process
(default: 30).}

\item{scan_rate}{The range within which to scan for optimizing tile positions
(default: 7).}
}
\value{
This function does not return a value; it generates a plot showing
the score optimization over iterations.
}
\description{
This function performs an optimization process to determine the optimal tile
positions in a gene sequence and plots the progression of the score
over iterations. It reads the overhang fidelity data, computes the optimal
number of tiles, and iteratively improves the positions, plotting the change
in score with respect to number of iterations at the end.
}
\examples{
execute_and_plot(max_iter = 5, scan_rate = 2)
}
\references{
Vladimir Potapov, Jennifer L. Ong, Rebecca B. Kucera, Bradley W. Langhorst,
Katharina Bilotti, John M. Pryor, Eric J. Cantor, Barry Canton,
Thomas F. Knight, Thomas C. Jr. Evans, and Gregory J. S. Lohman.
Comprehensive Profiling of Four Base Overhang Ligation Fidelity by
T4 DNA Ligase and Application to DNA Assembly.
ACS Synthetic Biology, 7(11):2665–2674, 2018.
Pryor JM, Potapov V, Kucera RB, Bilotti K, Cantor EJ, et al.
Enabling one-pot Golden Gate assemblies of unprecedented
complexity using data-optimized assembly design.
PLOS ONE 15(9): e0238592, 2020.
Alberts B., Johnson A., Lewis J., et al.
"Molecular Biology of the Cell." 6th edition.
Garland Science, 2014.
}
