% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{oligo_cost}
\alias{oligo_cost}
\title{Calculate Oligo Cost}
\usage{
oligo_cost(num_tiles, num_codons)
}
\arguments{
\item{num_tiles}{The number of tiles in the oligo sequence.}

\item{num_codons}{The number of codons in the oligo sequence.}
}
\value{
The calculated price per codon, considering both ccdB cost and
base replacement cost.
}
\description{
This function calculates the cost of oligonucleotides (oligos) based on
the number of tiles and the number of codons. It considers the cost of
ccdB (a cloning vector) and the cost of base replacement per tile.
}
\examples{
# Calculate oligo cost for 10 tiles and 100 codons
oligo_cost(10, 100)

}
