% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TRexDAD.R
\name{run_TRexDAD}
\alias{run_TRexDAD}
\title{Run Shiny App For Package TRexDAD}
\usage{
run_TRexDAD()
}
\value{
It opens the window for the shiny app.
}
\description{
A function that launches the shiny app for this package.
The shiny app offers users an interactive interface that allows them to
put in their gene of interest for assembly design.
Supplementary results, including the final positions, optimal tiles,
gene codons, oligo cost, overhang_fidelity chart,
are also provided to the user.
}
\examples{
\dontrun{
run_TRexDAD()
}

}
\references{
#' @references
Vladimir Potapov, Jennifer L. Ong, Rebecca B. Kucera, Bradley W. Langhorst,
Katharina Bilotti, John M. Pryor, Eric J. Cantor, Barry Canton,
Thomas F. Knight, Thomas C. Jr. Evans, and Gregory J. S. Lohman.
Comprehensive Profiling of Four Base Overhang Ligation Fidelity by
T4 DNA Ligase and Application to DNA Assembly.
ACS Synthetic Biology, 7(11):2665–2674, 2018.
Pryor JM, Potapov V, Kucera RB, Bilotti K, Cantor EJ, et al.
Enabling one-pot Golden Gate assemblies of unprecedented
complexity using data-optimized assembly design.
PLOS ONE 15(9): e0238592, 2020.
Alberts B., Johnson A., Lewis J., et al.
"Molecular Biology of the Cell." 6th edition.
Garland Science, 2014.
}
