% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{split_into_codons}
\alias{split_into_codons}
\title{Split Gene Sequence into Codons}
\usage{
split_into_codons(gene_sequence)
}
\arguments{
\item{gene_sequence}{A character string representing the gene sequence.}
}
\value{
A character vector where each element is a codon (a sequence of
three nucleotides), starting from the first nucleotide, with 'ATG'
as the second codon.
}
\description{
This function takes a gene sequence (as a string) and splits it into codons.
A codon consists of three nucleotides, and this function organizes the
sequence into these triplets. It verifies that the input is a valid DNA sequence,
that its length is a multiple of 3, contains 'ATG' as the second codon,
and has an extra codon both before the start and after the end codon of the ORF.
}
\examples{
# Example gene sequence
gene_seq <- "GCTATGGCCATTGTAATGGGCCGCTGAAAGGGTGCCCGATAGATC"
split_into_codons(gene_seq)

}
