% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{calculate_global_score}
\alias{calculate_global_score}
\title{Calculate Global Score for a List of Positions}
\usage{
calculate_global_score(gene_codons, tile_length, pos_lst)
}
\arguments{
\item{pos_lst}{A numeric vector representing positions in the gene sequence.}
}
\value{
The global score, which is a composite measure considering the off-target
reactions, repetitions in overhangs, and the sum of local scores for each position.
}
\description{
This function calculates a global score for a list of positions in a gene sequence.
The score takes into account off-target reactions, repetitions, and the sum of
local scores. It utilizes a pre-defined overhang fidelity dataframe for calculations.
}
