% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{calculate_optimal_tiles}
\alias{calculate_optimal_tiles}
\title{Calculate Optimal Number of Tiles and Cost}
\usage{
calculate_optimal_tiles(num_codons)
}
\arguments{
\item{num_codons}{An integer representing the total number of codons.}
}
\value{
A list containing three elements: \code{optimal_tiles}, the number of tiles
that minimize the oligo cost, \code{optimal_tile_length}, the average
length of these tiles, and \code{optimal_cost}, the cost associated with
the optimal number of tiles. Additionally, the function prints a
structured sentence summarizing these outputs.
}
\description{
This function determines the optimal number of tiles for a given number of
codons. It calculates the oligo cost for a range of tile numbers and finds
the number of tiles that minimizes this cost. In addition, the function computes
the optimal length of the tiles and the optimal cost associated with the optimal
number of tiles. The function limits the number of tiles to 50 at maximum, based
on the reference paper's indication that assembly fidelity decreases with an
increasing number of overhang pairs.
}
\examples{
# Calculate optimal tiles and cost for 100 codons
calculate_optimal_tiles(100)

}
\references{
Pryor, J. M. et al. Enabling one-pot Golden Gate assemblies of
unprecedented complexity using data-optimized assembly design. PLOS ONE 15,
e0238592 (2020).
}
