% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution_plot.R
\name{execute_and_plot}
\alias{execute_and_plot}
\title{Execute the Optimization Process and Plot Results}
\usage{
execute_and_plot(iteration_max = 10, scan_rate = 7)
}
\arguments{
\item{iteration_max}{The maximum number of iterations for the optimization process (default: 30).}

\item{scan_rate}{The range within which to scan for optimizing tile positions (default: 7).}
}
\value{
This function does not return a value; it generates a plot showing the score
optimization over iterations.
}
\description{
This function performs an optimization process to determine the optimal tile
positions in a gene sequence and plots the progression of the score over iterations.
It reads the overhang fidelity data, computes the optimal number of tiles, and iteratively
improves the positions, plotting the score at each iteration.
}
