% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{get_all_overhangs}
\alias{get_all_overhangs}
\title{Get All Overhang Sequences for a List of Positions}
\usage{
get_all_overhangs(gene_codons, pos_lst)
}
\arguments{
\item{gene_codons}{A list of splitted codons of target genes}

\item{pos_lst}{A numeric vector representing positions in the gene sequence. Each element in pos_lst is the start of a gene tile.}
}
\value{
A list of overhang sequences corresponding to each position in
\code{pos_lst}. Each position contributes two elements to the list: the
head and tail overhangs.
}
\description{
This function computes overhang sequences for a given list of positions
in a gene sequence. It iteratively calls the \code{get_overhangs} function to
calculate the head and tail overhangs for each position and accumulates
them in a list.
}
