% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{get_overhangs}
\alias{get_overhangs}
\title{Get Overhang Sequences for Single Tile}
\usage{
get_overhangs(gene_codons, start, end, as_dna_strings)
}
\arguments{
\item{gene_codons}{A list of splitted codons of target genes}

\item{start}{An integer indicating the start position of the tile.}

\item{end}{An integer indicating the end position of the tile.}

\item{as_dna_strings}{A boolean indicating the return type of the sequences.
If TRUE, sequences are returned as \code{DNAString} objects;
if FALSE, as character strings.}
}
\value{
A list containing two elements: \code{head} and \code{tail}, representing the
overhang sequences at the specified positions.
}
\description{
This function calculates the overhang sequences at the specified positions
in a gene sequence. Depending on the flag, it returns these sequences as
either \code{DNAString} objects or plain character strings.
}
\examples{
# Overhangs for sample gene_codons at position(start, end)
get_overhangs(c("GAG", "CTG", "TGT", "AGG", "TGC", "CGG", "CCA", 
"ATT", "TGA", "TAG", "GAA", "TAT", "AGC"), 3, 4, TRUE)
}
