% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pick_position}
\alias{pick_position}
\title{Pick a Position for Optimization}
\usage{
pick_position(gene_codons, tile_length, pos_lst)
}
\arguments{
\item{gene_codons}{A vector of DNA codons.}

\item{tile_length}{An integer representing the expected length of the tile.}

\item{pos_lst}{A numeric vector representing positions in a gene sequence.}
}
\value{
A list with two elements: \code{index}, the index in \code{pos_lst} of the picked
position (shifted by 1), and \code{position}, the actual value of the picked position.
}
\description{
This function selects a position from a list of positions for optimization
based on their scores. Positions with lower scores are more likely to be picked.
It uses a weighted sampling approach where the weights are inversely proportional
to the scores of the positions.
}
\examples{
gene_codons <- c("GAG", "ATG", "TGT", "AGG", "TGC", "CGG", "CCA", 
                  "ATT", "TGA", "TAG", "GAA", "TGA", "AGC")
tile_length <- 5
pos_lst <- c(3, 6, 8, 11)
pick_position(gene_codons, tile_length, pos_lst)
}
