% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{split_into_codons}
\alias{split_into_codons}
\title{Split Gene Sequence into Codons}
\usage{
split_into_codons(gene_sequence)
}
\arguments{
\item{gene_sequence}{A character string representing the gene sequence.}
}
\value{
A character vector where each element is a codon (a sequence of
three nucleotides).
}
\description{
This function takes a gene sequence (as a string) and splits it into codons.
A codon consists of three nucleotides, and this function organizes the
sequence into these triplets.
}
\examples{
# Example gene sequence
gene_seq <- "ATGGCCATTGTAATGGGCCGCTGAAAGGGTGCCCGATAG"
split_into_codons(gene_seq)

}
